+++ presentation/src/main/java/com/threegap/bitnagil/presentation/setting/SettingViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.threegap.bitnagil.domain.auth.usecase.LogoutUseCase
 import com.threegap.bitnagil.presentation.common.version.VersionNameProvider
-import com.threegap.bitnagil.presentation.setting.model.mvi.SettingSideEffect
-import com.threegap.bitnagil.presentation.setting.model.mvi.SettingState
+import com.threegap.bitnagil.presentation.setting.contract.SettingSideEffect
+import com.threegap.bitnagil.presentation.setting.contract.SettingState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
