+++ presentation/src/main/java/com/threegap/bitnagil/presentation/setting/contract/SettingSideEffect.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.presentation.setting.contract
+
+sealed interface SettingSideEffect {
+    data object NavigateToLogin : SettingSideEffect
+    data object NavigateToWithdrawal : SettingSideEffect
+}
