--- presentation/src/main/java/com/threegap/bitnagil/presentation/setting/model/mvi/SettingSideEffect.kt
@@ -1,6 +1,0 @@
-package com.threegap.bitnagil.presentation.setting.model.mvi
-
-sealed class SettingSideEffect {
-    data object NavigateToLogin : SettingSideEffect()
-    data object NavigateToWithdrawal : SettingSideEffect()
-}
